Update StandFunctionTrade path, and fix some bugs on get products request
