+++ frontend/src/pages/workspace/StandFunctionTrade/FormTrade/FormTrade.module.scss
