+++ frontend/src/pages/workspace/StandFunctionTrade/LastTradeList/LastPurchaseList.module.scss
@@ -1,0 +1,41 @@
+@use "@styles/variables" as *;
+
+.body {
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  width: 80vw;
+  max-width: 1200px;
+  z-index: 100;
+  position: fixed;
+  background-color: $color-gray-100;
+  padding: 10px;
+  border-radius: $border-radius;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  .mainList {
+    flex: 1;
+    width: 100%;
+    padding: 2px 4px;
+    border-radius: $border-radius;
+    background-color: $color-gray-000;
+    box-sizing: border-box;
+
+    .header {
+      background-color: $color-primary-2;
+    }
+
+    li {
+      display: grid;
+      grid-template-columns: 1fr 1fr;
+      grid-auto-flow: row;
+      gap: 2px;
+      margin-bottom: 6px;
+      padding: 2px 6px;
+      border-radius: $border-radius;
+      background-color: $color-primary-1;
+    }
+  }
+}
