+++ frontend/src/pages/workspace/StandFunctionTrade/LastTradeList/ShowLastTrade/ShowLastPurchase.module.scss
@@ -1,0 +1,66 @@
+@use "@styles/variables" as *;
+
+.body {
+  width: 100%;
+  margin: 6px 0;
+  padding: 6px;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  align-self: center;
+  background-color: $color-gray-000;
+  border-radius: $border-radius;
+
+  p {
+    text-align: center;
+  }
+
+  h4 {
+    font-weight: 500;
+    text-align: center;
+  }
+
+  .timestamp {
+    padding: 2px;
+    font-weight: 300;
+  }
+
+  ul {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    .listHeader {
+      background-color: $color-primary-2;
+    }
+
+    li {
+      margin-bottom: 4px;
+      padding: 2px 4px;
+      width: 100%;
+      box-sizing: border-box;
+      border-radius: $border-radius;
+      display: grid;
+      grid-template-columns: 1fr 1fr 1fr;
+      grid-auto-flow: row;
+      gap: 2px;
+      background-color: $color-primary-1;
+    }
+  }
+
+  .delete {
+    display: flex;
+    justify-content: center;
+    padding: 6px 0 4px;
+
+    button {
+      padding: 4px 12px;
+      background-color: $color-primary-2;
+
+      p {
+        font-size: 1rem;
+        font-weight: 500;
+      }
+    }
+  }
+}
