+++ frontend/src/pages/workspace/StandFunctionTrade/StandFunctionTrade.module.scss
